+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v2.1.0
+## New
+- Added support for Meilisearch v1.2
+- Added `/import/cocktail` endpoint
+    - Supports importing cocktails in YAML and JSON formats
+    - Supports importing cocktails from URLs that are supported by built in scrapers
+- Added `/cocktail/{id}/share` endpoint
+    - Supports sharing cocktails in JSON, YAML, XML and plain text format
+
+## Changes
+- Disabled foreign key checks when importing data from zip file #117
+
 # v2.0.0
 ## Breaking changes
 - Minimum supported Meilisearch version is 1.1